Refactor MOUNT and IMGMOUNT out of dos_programs
